Added weather display + refactor Allinonefortasks.js to modules.
