+++ newTask.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://kit.fontawesome.com/602ba1bef5.js" crossorigin="anonymous"></script>
     <link rel="stylesheet" href="">
-    <script src="./src/components/Allinonefortasks.js" type="module" defer></script>
+    
     <title>My List</title>
 </head>
 <body>
@@ -48,6 +48,6 @@
             </table>
     </div>
 
-
+    <script src="./src/components/Allinonefortasks.js" type="module" defer></script>
 </body>
 </html>
